web-app/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-environ==0.10.0
django-filter==22.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
openpyxl==3.1.1
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pydantic==1.10.5
pytest==7.2.2
pytest-django==4.5.2
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
celery[redis]==5.2.7 